move metrics into their own module
